Add option to encode state as base64 when sending to oauth providers
Some providers do not like the default encoding for state strings.

To work around, if the oauth section of the provider contains:

```
state_encoding: 'base64'
```

Then we will encode the state string as base64 before sending off to the
upstream provider.  In redirect.html, if the string starts with eyJ,
then we reverse the encoding.